Move basecommand to pip._internal.cli
