+++ src/pip/_internal/cli/basecommand.py
