+++ src/pip/_internal/commands/download.py
@@ -4,7 +4,7 @@ import logging
 import os
 
 from pip._internal import cmdoptions
-from pip._internal.basecommand import RequirementCommand
+from pip._internal.cli.basecommand import RequirementCommand
 from pip._internal.exceptions import CommandError
 from pip._internal.index import FormatControl
 from pip._internal.operations.prepare import RequirementPreparer
