+++ src/pip/_internal/commands/freeze.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 import sys
 
 from pip._internal import index
-from pip._internal.basecommand import Command
+from pip._internal.cli.basecommand import Command
 from pip._internal.cache import WheelCache
 from pip._internal.compat import stdlib_pkgs
 from pip._internal.operations.freeze import freeze
