+++ src/pip/_internal/commands/list.py
@@ -6,7 +6,7 @@ import logging
 from pip._vendor import six
 from pip._vendor.six.moves import zip_longest
 
-from pip._internal.basecommand import Command
+from pip._internal.cli.basecommand import Command
 from pip._internal.cmdoptions import index_group, make_option_group
 from pip._internal.exceptions import CommandError
 from pip._internal.index import PackageFinder
