+++ src/pip/_internal/commands/show.py
@@ -7,7 +7,7 @@ from email.parser import FeedParser  # type: ignore
 from pip._vendor import pkg_resources
 from pip._vendor.packaging.utils import canonicalize_name
 
-from pip._internal.basecommand import Command
+from pip._internal.cli.basecommand import Command
 from pip._internal.status_codes import ERROR, SUCCESS
 
 logger = logging.getLogger(__name__)
